pyproject.toml
[project]
name = "authz"
version = "0.1.0"
description = ""
authors = [
    {name = "Chas Sweeting",email = "77371250+clsweeting@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "cachetools (>=5.5.2,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

